requirements.txt
pandas
requests 
beautifulsoup4
SQLAlchemy
tqdm
psycopg2-binary
python-dotenv
sqlmodel